add a list of user-defined boards from arduinodir (temp)
In future, instead of directly editing the arduino directory's boards.txt
we should have arduino.mk search for boards first in our user boards.txt
file, then in the official arduino boards.txt file.

In line with our goal of not requiring the user to delve into the arduino
directory and change things.