[sdbrowser] add nano, rm and cl fns, fix line overflow bug
nano: create a new file and write one line (ended by a '\r') to it
rm: remove a file
cl: print 50 newlines to serial

line overflow bugfix:
previously, if we exceeded the line limit in one command, sdbrowser
would complain, not act on the partial command, and clear the line
buffer. However, it would fail to reset the error flag to zero, so
when we next entered a line that falls within the char limit,
sdbrowser would (wrongly) complain that it is too long.