[gpssaver] GPS.read() in loop, parse/print, save raw
Previously we read the GPS using a timer interrupt, but switching
to a loop-based method made printing the raw data better as the
problem of incomplete lines being printed went away

Demonstrates the printing of parsed GPS data using the Adafruit
parser.

Raw data is being saved to the SD card.