[cheesesock, csockr] Trim down CheeseSock, write example
I removed a lot of the cruft from CheeseSock that was
SnS-related. I modified it to use only SoftwareSerial.
I removed the singleton pattern and used the style of
the Adafruit GPS code - no more caling getInstance().
I removed the registerAppFn and we can now only register
non-member functions. Lots of changes but the basic
functionality remains.

csockr is example code. Demonstrates escaping of slash.