[clctr, gpssaver] allow request of saved packets
Had to implement SoftwareSerial listen() switching in
order to receive from two SoftwareSerial ports (GPS
and CheeseSock).

Reference:
http://arduiniana.org/libraries/newsoftserial/

Implemented block requests. The clctr arduino can
send a request packet with the start and end indices
of the packet range that it wants to receive from the
gpssaver.

Currently block requests with size up to 1000 have
been tested, with the GPS new packets enabled and
disabled. When the new packets from the GPS (those
that show the current location) are not sent, and
only the old packets are transmitted, we have full
packet delivery at a delay of 5ms between each.

When GPS is enabled (so the wheelchair is trying to
send current data as well as backed-up data), some
packets get lost (103 out of 107 delivered). Suspect
this happens when the new packet and the old packet
arrive at the netmod arduino at roughly the same time.

Could be investigated further but prob. not worth it.