[clctr] enable pretty-printing and PWM ctrl
When the GPS has no fix, the wheelchair still
sends out packets, containing null data for the
timestamp/location and real data for tilt and
bumpiness, since the accelerometer still works.
Such packets are given an id of 9999, and clctr
should print them in a pretty manner. This wasn't
used during the demo though.

The important part is the PWM control of the LED
and ammeter to show bumpiness and tilt. This was
used during the Q&A.