[netmod] host accelerometer, allow demo mode
Now, netmod is responsible for polling the
accelerometer and calculating the average tilt
over the period in between the requests from
the gpssaver for data.

Ditto for variance in the vertical direction.
Algorithm for variance is from Knuth (wiki)

For demo purposes we can set pin 3 high and
spur the netmod to pump out packets containing
tilt and bumpiness data at a high rate. This
allows the ammeter-needle to move smoothly as
the judges play with the tilted box.