[testecho] test program to debug serial garbling at 115200
Somehow, using the arduino-serial-lib.c at baud rate
115200 results in errors. If you try to send something
using the arduino-serial-lib.c to an Arduino running this
testecho program, you will find that what you send is not
the same as what you receive.

The solution to this problem was to halve the baud rate
of the serial comms to 57600, in the clctr. Because only
the clctr needs to interact with a arduino-serial-lib.c
program (the C program running on the base station)